init changes

+++ src/app/dialogs/delete/delete.dialog.component.ts
@@ -18,6 +18,7 @@ export class DeleteDialogComponent {
   }
 
   confirmDelete(): void {
+    if(this.data)
     this.dataService.deleteIssue(this.data.invoice_no);
   }
 }
